+++ src/main/java/edu/brown/cs/stormrouter/route/PathWeatherInfo.java
@@ -32,12 +32,12 @@ public class PathWeatherInfo {
    * 
    * @param pointLat    latitude of given point
    * @param pointLong   longitude of given point
-   * @param weatherType Stringr representing type of weather to be displayed
+   * @param weatherType String representing type of weather to be displayed
    * @param weatherSum  String summary of weather occurring at this point
    * @param scoreIncr   integer representing score of the given weather point
    * @param pointTime   Unix time at which point is reached
    */
-  public void addWeatherData(float pointLat, float pointLong,
+  public void addWeatherData(double pointLat, double pointLong,
       String weatherType, String weatherSum, int scoreIncr, long pointTime) {
     // Adds details to set
     weatherData.add(new PathWeatherPoint(pointLat, pointLong, weatherType,
